chore: add panes to embed and some refactors
